Eject v2 plugins into its own workspace
With this change, v2 plugins reside in a separate workspace, and
they no longer have to be explicitly listed in the package.json of
`discourse-plugin-v2`. Instead, it discovers the plugins during the
build and uses a custom/modified node resolution process to resolve
their modules.

Anything that is not an Ember implicit module (defined by Embroider
via the `addon.dependencies()` rollup plugin) and not listed as a
`peerDependencies` of `discourse-plugin-v2` will be _bundled_ as
part of the build.
