+++ app/assets/javascripts/discourse-plugins-v2/.gitignore
@@ -1,0 +1,1 @@
+src
