+++ app/assets/javascripts/package.json
@@ -6,7 +6,6 @@
   "workspaces": [
     "admin",
     "bootstrap-json",
-    "chat-lite",
     "deprecation-silencer",
     "dialog-holder",
     "discourse",
@@ -15,7 +14,6 @@
     "discourse-i18n",
     "discourse-markdown-it",
     "discourse-plugin",
-    "discourse-plugin-dev",
     "discourse-plugins",
     "discourse-plugins-v2",
     "discourse-widget-hbs",
