+++ plugins-v2/chat-lite/addon-main.cjs
