+++ plugins-v2/chat-lite/babel.config.json
