+++ plugins-v2/chat-lite/rollup.config.mjs
