+++ plugins-v2/chat-lite/src/components/header-icon.gjs
