+++ plugins-v2/chat-lite/src/connectors/extra-header-icons.gjs
