+++ plugins-v2/chat-lite/src/markdown-features/scramble.js
