+++ plugins-v2/discourse-plugin-dev/package.json
