+++ plugins-v2/discourse-plugin-dev/src/rollup/export-plugin-features.js
