+++ plugins-v2/discourse-plugin-dev/src/rollup/index.js
