+++ plugins-v2/package.json
@@ -1,0 +1,6 @@
+{
+  "private": true,
+  "workspaces": [
+    "*"
+  ]
+}
